[DEV-3310] Split IDVs from Awards
