+++ usaspending_api/api_contracts/contracts/v2/idvs/funding_rollup.md
@@ -1,7 +1,7 @@
 FORMAT: 1A
 HOST: https://api.usaspending.gov
 
-# IDV Funding Rollup [/api/v2/awards/idvs/funding_rollup/]
+# IDV Funding Rollup [/api/v2/idvs/funding_rollup/]
 
 This end point returns award metadata summing the total transaction obligations and counting awarding agencies, funding agencies, and federal accounts for an IDV's children and grandchildren.
