+++ usaspending_api/idvs/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class IDVsConfig(AppConfig):
+    name = "idvs"
