+++ usaspending_api/idvs/v2/views/funding_rollup.py
@@ -64,7 +64,7 @@ class IDVFundingRollupViewSet(APIView):
     """
     Returns File C funding totals associated with an IDV's children.
     """
-    endpoint_doc = "usaspending_api/api_docs/api_documentation/awards/idvs/funding_rollup.md"
+    endpoint_doc = "usaspending_api/api_contracts/contracts/v2/idvs/funding_rollup.md"
 
     @staticmethod
     def _business_logic(request_data: dict) -> OrderedDict:
