+++ usaspending_api/urls.py
@@ -32,7 +32,7 @@ urlpatterns = [
     url(r"^api/v1/transactions/", include("usaspending_api.awards.v1.urls_transactions")),
     url(r"^api/v2/autocomplete/", include("usaspending_api.references.v2.urls_autocomplete")),
     url(r"^api/v2/awards/", include("usaspending_api.awards.v2.urls_awards")),
-    url(r"^api/v2/awards/idvs/", include("usaspending_api.awards.v2.urls_idv_awards")),
+    url(r"^api/v2/idvs/", include("usaspending_api.idvs.v2.urls_idvs")),
     url(r"^api/v2/award_spending/", include("usaspending_api.awards.v2.urls_award_spending")),
     url(r"^api/v2/subawards/", include("usaspending_api.awards.v2.urls_subawards")),
     url(r"^api/v2/budget_authority/", include("usaspending_api.accounts.urls_budget_authority")),
